pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Stanislav Buldyhin <faquedust@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
channels-redis = "^4.1.0"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
pyjwt = "^2.8.0"
python-json-logger = "^2.0.7"
gunicorn = "^21.2.0"
uvicorn = { extras = ["standard"], version = "^0.25.0" }
channels = "^4.0.0"
whitenoise = "^6.6.0"
psycopg = { extras = ["binary"], version = "^3.1.16" }
strawberry-graphql = {extras = ["channels"], version = "^0.217.0"}


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
black = "^23.12.0"
factory-boy = "^3.3.0"
coverage = "^7.3.3"
isort = "^5.13.2"
daphne = "^4.0.0"
flake8 = "^6.1.0"
flake8-isort = "^6.1.1"
radon = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py312']
line-length = 120
exclude = '''
(
    \.venv
    | .+/migrations
)
'''

[tool.coverage.report]
exclude_also = ["CHANNEL_LAYERS"]
omit = ["manage.py", "core/testrunner.py", "core/settings.py"]

[tool.isort]
profile = "black"
force_single_line = true
order_by_type = false
line_length = 120
skip = '.venv,migrations'
